requirements.txt
torch
numpy
colorama
pillow